restructured application
