+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "lok-buch"
-version = "0.2.0"
+version = "0.3.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
