+++ src/app/backend/test/util.rs
@@ -1,4 +1,4 @@
-use crate::backend::database::lok::Lok;
+use crate::app::backend::database::lok::Lok;
 use crate::database::sqlite_db::SQLiteDB;
 use crate::database::Database;
 use async_std::task;
