+++ src/app/page/loading_page.rs
@@ -1,7 +1,7 @@
 use crate::app::message::Message;
 use crate::app::page::{Page, Pages};
+use crate::app::ui;
 use crate::app::Lokbuch;
-use crate::ui;
 use iced::widget::{center, text};
 use iced::{Element, Task};
