+++ src/app/state.rs
@@ -1,5 +1,5 @@
+use crate::app::backend::database::lok::Lok;
 use crate::app::Message;
-use crate::backend::database::lok::Lok;
 use iced::Task;
 
 /// The State holds all data that is needed during runtime.
