+++ src/app/stored_data.rs
@@ -1,5 +1,5 @@
-use crate::backend::resource_manager::LokResourceManager;
-use crate::backend::sqlite_backend::SQLiteBackend;
+use crate::app::backend::resource_manager::LokResourceManager;
+use crate::app::backend::sqlite_backend::SQLiteBackend;
 
 #[derive(Clone, Debug)]
 pub struct StoredData {
