+++ src/app/ui.rs
@@ -10,5 +10,5 @@ pub const NO_DATA_AVAILABLE_TEXT: &str = "---";
 pub const ICON_PATH: &str = "res/images/icon.png";
 
 pub fn moving_icon_frames() -> iced_gif::Frames {
-    iced_gif::Frames::from_bytes(include_bytes!("../res/images/icon_move.gif").into()).expect("Decoding gif failed!")
+    iced_gif::Frames::from_bytes(include_bytes!("../../res/images/icon_move.gif").into()).expect("Decoding gif failed!")
 }
